// src/main/webapp/app/shared/lookup/lookup.enum.ts
export enum LookupEnum {
  FarmType = 'FARM_TYPE',
  PigeonStatus = 'PIGEON_STATUS',
  EggStatus = 'EGG_STATUS',
  ChickStatus = 'Chick_STATUS',
  TreatmentType = 'TREATMENT_TYPE',
  IncomeType = 'INCOME_TYPE',
  ExpenseType = 'EXPENSE_TYPE',
  InvoiceStatus = 'INVOICE_STATUS',
  SubscriptionStatus = 'SUBSCRIPTION_STATUS',
  SheepCategory = 'SHEEP_CATEGORY',
  SheepType = 'SHEEP_TYPE',
  SheepStatus = 'SHEEP_STATUS',
  SheepGender = 'SHEEP_GENDER',
  GestationStatus = 'GESTATION_STATUS',
  SheepVaccinationType = 'SHEEP_VACCINATION_TYPE',
  SheepTreatmentType = 'SHEEP_TREATMENT_TYPE',
  SheepTreatmentDoseType = 'SHEEP_TREATMENT_DOSE_TYPE',
}